move project and notebook actions back to SamAuthProvider
