+++ src/main/scala/org/broadinstitute/dsde/workbench/leonardo/auth/SwaggerSamClient.scala
@@ -4,10 +4,8 @@ import java.io.{ByteArrayInputStream, File}
 import java.util.concurrent.TimeUnit
 
 import akka.http.scaladsl.model.StatusCodes
-import net.ceedubs.ficus.Ficus._
 import com.google.api.client.googleapis.auth.oauth2.GoogleCredential
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
-import com.google.api.client.googleapis.json.GoogleJsonError
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.services.plus.PlusScopes
 import com.google.auth.oauth2.ServiceAccountCredentials
@@ -16,18 +14,13 @@ import com.google.gson.Gson
 import com.google.gson.internal.LinkedTreeMap
 import io.swagger.client.ApiClient
 import io.swagger.client.api.{GoogleApi, ResourcesApi}
-import org.broadinstitute.dsde.workbench.leonardo.model.Actions.Action
 import org.broadinstitute.dsde.workbench.leonardo.model.{ClusterName, LeoException}
-import org.broadinstitute.dsde.workbench.leonardo.model.NotebookClusterActions._
-import org.broadinstitute.dsde.workbench.leonardo.model.ProjectActions.{CreateClusters, ProjectAction}
 import org.broadinstitute.dsde.workbench.model.WorkbenchEmail
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future}
 
-case class NotebookActionError(action: Action) extends
-  LeoException(s"${action.toString} was not recognized", StatusCodes.NotFound)
 
 case class UserEmailAndProject(userEmail: WorkbenchEmail, googleProject: GoogleProject)
 
@@ -107,29 +100,6 @@ class SwaggerSamClient(samBasePath: String, cacheExpiryTime: Int, cacheMaxSize: 
     credential.refreshAccessToken.getTokenValue
   }
 
-  //gets the string we want for each type of action - definitely NOT how we want to do this in the long run
-  protected def getProjectActionString(action: Action): String = {
-    projectActionMap.getOrElse(action, throw NotebookActionError(action))
-  }
-
-  //gets the string we want for each type of action - definitely NOT how we want to do this in the long run
-  protected def getNotebookActionString(action: Action): String = {
-    notebookActionMap.getOrElse(action, throw NotebookActionError(action))
-  }
-
-  private def projectActionMap: Map[Action, String] = Map(
-    //list_notebook_cluster
-    CreateClusters -> "launch_notebook_cluster",
-    SyncDataToCluster -> "sync_notebook_cluster",
-    DeleteCluster -> "delete_notebook_cluster")
-
-
-  private def notebookActionMap: Map[Action, String] = Map(
-    CreateClusters -> "launch_notebook_cluster",
-    GetClusterStatus -> "status",
-    ConnectToCluster -> "connect",
-    SyncDataToCluster -> "sync",
-    DeleteCluster -> "delete")
 
   private def getClusterResourceId(googleProject: GoogleProject, clusterName: ClusterName): String = {
     googleProject.value + "_" + clusterName
@@ -144,12 +114,12 @@ class SwaggerSamClient(samBasePath: String, cacheExpiryTime: Int, cacheMaxSize: 
     resourcesApiAsPet(userEmail, googleProject).deleteResource(notebookClusterResourceTypeName, getClusterResourceId(googleProject, clusterName))
   }
 
-  def hasActionOnBillingProjectResource(userEmail: WorkbenchEmail, googleProject: GoogleProject, action: Action): Boolean = {
-    hasActionOnResource(billingProjectResourceTypeName, googleProject.value, userEmail, googleProject, getProjectActionString(action))
+  def hasActionOnBillingProjectResource(userEmail: WorkbenchEmail, googleProject: GoogleProject, action: String)(implicit executionContext: ExecutionContext): Boolean = {
+    hasActionOnResource(billingProjectResourceTypeName, googleProject.value, userEmail, googleProject, action)
   }
 
-  def hasActionOnNotebookClusterResource(userEmail: WorkbenchEmail, googleProject: GoogleProject, clusterName: ClusterName, action: Action): Boolean = {
-    hasActionOnResource(notebookClusterResourceTypeName, getClusterResourceId(googleProject, clusterName), userEmail, googleProject, getNotebookActionString(action))
+  def hasActionOnNotebookClusterResource(userEmail: WorkbenchEmail, googleProject: GoogleProject, clusterName: ClusterName, action: String)(implicit executionContext: ExecutionContext): Boolean = {
+    hasActionOnResource(notebookClusterResourceTypeName, getClusterResourceId(googleProject, clusterName), userEmail, googleProject, action)
   }
 
   private def hasActionOnResource(resourceType: String, resourceName: String, userEmail: WorkbenchEmail, googleProject: GoogleProject, action: String)(implicit executionContext: ExecutionContext): Boolean = {
